Refactor directory structure for modern packaging standards and adjust tests accordingly.
