+++ pyproject.toml
@@ -2,3 +2,35 @@
 requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
+[project]
+name = "amisrsynthdata"
+version = "0.0.1"
+authors = [
+  { name="L. Lamarche", email="leslie.lamarche@sri.com" },
+]
+description = "Tool for generating synthetic AMISR data files."
+readme = "README.rst"
+license = { file = "LICENSE" }
+keywords = ["AMISR", "synthetic", "data"]
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Topic :: Scientific/Engineering",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: GNU General Public License (GPL)",
+    "Natural Language :: English",
+    "Programming Language :: Python",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/amisr/amisrsynthdata"
+#"Bug Tracker" = "https://github.com/pypa/sampleproject/issues"
+
+[project.scripts]
+    amisrsynthdata = "amisrsynthdata.syntheticdata:main"
+
+[tool.pytest.ini_options]
+addopts = [
+    "--import-mode=importlib",
+]
