--- setup.cfg
@@ -1,35 +1,0 @@
-[metadata]
-name = amisrsynthdata
-version = attr: amisrsynthdata.__version__
-description = Tool for generating synthetic AMISR data files.
-long_description = file: README.rst, LICENSE.rst
-keywords = AMISR, synthetic, data
-license = GNU GPLv3
-author = L. Lamarche
-author_email = leslie.lamarche@sri.com
-url = https://github.com/amisr/amisrsynthdata
-classifiers =
-    Programming Language :: Python :: 3
-    Topic :: Scientific/Engineering
-    Intended Audience :: Science/Research
-    License :: OSI Approved :: GNU General Public License (GPL)
-    Natural Language :: English
-    Programming Language :: Python
-
-
-[options]
-packages = find:
-install_requires =
-    numpy
-    h5py
-    pymap3d
-    pyyaml
-    importlib_resources
-
-[options.entry_points]
-console_scripts =
-    amisrsynthdata = amisrsynthdata.syntheticdata:main
-
-[options.package_data]
-amisrsynthdata =
-  beamcodes/*.txt
