+++ src/amisrsynthdata/beamcodes/bcotable_pfisr.txt
