+++ src/amisrsynthdata/beamcodes/bcotable_risrc.txt
