+++ src/amisrsynthdata/beamcodes/bcotable_risrn.txt
