+++ src/amisrsynthdata/state_functions/density.py
