+++ src/amisrsynthdata/state_functions/utils.py
