+++ tests/test_Ionosphere.py
@@ -5,13 +5,14 @@ Test of amisrsynthdata Ionosphere class
 import pytest
 #import amisrsynthdata
 import yaml
+import os
 import numpy as np
 from apexpy import Apex
-from src.ionosphere import Ionosphere
+from amisrsynthdata.ionosphere import Ionosphere
 
 @pytest.fixture
 def config():
-    config_file = 'config.yaml'
+    config_file = os.path.join(os.path.dirname(__file__), 'config.yaml')
     with open(config_file, 'r') as cf:
         config = yaml.load(cf, Loader=yaml.FullLoader)
     return config
@@ -66,7 +67,7 @@ def test_itemp(ionosphere, config, varin):
 def test_sum_density(config, varin):
 
     config['DENSITY'].append({'uniform': {'value': 5.e10}})
-    iono = amisrsynthdata.Ionosphere(config)
+    iono = Ionosphere(config)
     dens = iono.density(*varin)
     truth_dens = config['DENSITY'][0]['uniform']['value'] + 5.e10
 
@@ -75,7 +76,7 @@ def test_sum_density(config, varin):
 def test_sum_velocity(config, varin):
 
     config['VELOCITY'].append({'uniform_glat_aligned': {'value': [200., 0., 0]}})
-    iono = amisrsynthdata.Ionosphere(config)
+    iono = Ionosphere(config)
     vel = iono.velocity(*varin)
     truth_vel = np.array(config['VELOCITY'][0]['uniform_glat_aligned']['value']) + np.array([200., 0., 0.])
     
@@ -84,7 +85,7 @@ def test_sum_velocity(config, varin):
 def test_sum_etemp(config, varin):
 
     config['ETEMP'].append({'uniform': {'value': 2000.}})
-    iono = amisrsynthdata.Ionosphere(config)
+    iono = Ionosphere(config)
     etemp = iono.etemp(*varin)
     truth_etemp = config['ETEMP'][0]['uniform']['value'] + 2000.
 
@@ -93,7 +94,7 @@ def test_sum_etemp(config, varin):
 def test_sum_itemp(config, varin):
 
     config['ITEMP'].append({'uniform': {'value': 2000.}})
-    iono = amisrsynthdata.Ionosphere(config)
+    iono = Ionosphere(config)
     itemp = iono.itemp(*varin)
     truth_itemp = config['ITEMP'][0]['uniform']['value'] + 2000.
