+++ tests/test_SyntheticData.py
@@ -11,11 +11,13 @@ import datetime as dt
 import filecmp
 import os
 
-from src.sytheticdata import SyntheticData
+from amisrsynthdata.syntheticdata import SyntheticData
+from amisrsynthdata.ionosphere import Ionosphere
+from amisrsynthdata.radar import Radar
 
 @pytest.fixture
 def config():
-    config_file = 'config.yaml'
+    config_file = os.path.join(os.path.dirname(__file__), 'config.yaml')
     with open(config_file, 'r') as cf:
         config = yaml.load(cf, Loader=yaml.FullLoader)
     return config
@@ -29,15 +31,15 @@ def synthdata(config):
 @pytest.fixture
 def datafile():
     
-    filename = 'synthetic_data.h5'
+    filename = os.path.join(os.path.dirname(__file__), 'synthetic_data.h5')
     h5file = h5py.File(filename, 'r')
     
     return h5file
 
 
 def test_init(synthdata):
-    assert isinstance(synthdata.radar, amisrsynthdata.Radar)
-    assert isinstance(synthdata.iono, amisrsynthdata.Ionosphere)
+    assert isinstance(synthdata.radar, Radar)
+    assert isinstance(synthdata.iono, Ionosphere)
     assert isinstance(synthdata.utime, np.ndarray)
     assert isinstance(synthdata.time, np.ndarray)
     assert isinstance(synthdata.ne, np.ndarray)
@@ -140,6 +142,11 @@ def hdf52dict(h5):
             continue
     return out
 
+def assert_dict_equal(dict1, dict2):
+    np.testing.assert_equal(dict1.keys(), dict2.keys())
+    for k in dict1.keys():
+        np.testing.assert_allclose(dict1[k], dict2[k])
+
 
 def test_generate_beamcodes(synthdata, datafile):
 
@@ -158,31 +165,32 @@ def test_generate_fitted_params(synthdata, datafile):
     truth_fit_info = hdf52dict(datafile['FittedParams/FitInfo'])
     truth_ne_from_power = hdf52dict(datafile['NeFromPower'])
 
-    np.testing.assert_equal(fitted_params, truth_fitted_params)
-    np.testing.assert_equal(fit_info, truth_fit_info)
-    np.testing.assert_equal(ne_from_power, truth_ne_from_power)
+    assert_dict_equal(fitted_params, truth_fitted_params)
+    assert_dict_equal(fit_info, truth_fit_info)
+    assert_dict_equal(ne_from_power, truth_ne_from_power)
 
 def test_generate_time(synthdata, datafile):
 
     time = synthdata.generate_time()
     truth_time = hdf52dict(datafile['Time'])
-    np.testing.assert_equal(time, truth_time)
+    assert_dict_equal(time, truth_time)
 
 def test_generate_geomag(synthdata, datafile):
     geomag = synthdata.generate_geomag()
     truth_geomag = hdf52dict(datafile['Geomag'])
-    np.testing.assert_equal(geomag, truth_geomag)
+    assert_dict_equal(geomag, truth_geomag)
 
 def test_generate_site(synthdata, datafile):
     site = synthdata.generate_site()
     truth_site = hdf52dict(datafile['Site'])
     # Remove 'Name' - not generated by function
     del truth_site['Name']
-    np.testing.assert_equal(site, truth_site)
+    assert_dict_equal(site, truth_site)
 
 def test_create_hdf5_output(synthdata):
     synthdata.create_hdf5_output('temp_out.h5')
-    assert filecmp.cmp('temp_out.h5', 'synthetic_data.h5')
+    truth_file = os.path.join(os.path.dirname(__file__), 'synthetic_data.h5')
+    assert filecmp.cmp('temp_out.h5', truth_file)
     os.remove('temp_out.h5')
 
 def test_create_summary_plots(synthdata, config):
